Revamp main menu, input routing, and hitbox API
Rework main menu UI to use a background sprite and transparent text buttons, add new menu assets, and centralize sprite keys in LoadSave. Route mouse events from Mouse to menu/play states and standardize mouse callback name from mouseRelease to mouseReleased (updated StateMethods, PlayStates, States, Mouse). Adjust Entity.initHitBox signature to use ints and update Player to use GameCore.SCALE for speed and hitbox dimensions. Also update MainMenuButton to expect three button states and reference the new MENU_BUTTONS constant. Add fallback exit handling in GameCore and include two new resource images.

+++ src/entity/Entity.java
@@ -25,7 +25,7 @@ public abstract class Entity {
 		g.drawRect((int) hitBox.x, (int) hitBox.y, (int) hitBox.width, (int) hitBox.height);
 	}
 
-	protected void initHitBox(float x, float y, float widht, float height) {
+	protected void initHitBox(float x, float y, int widht, int height) {
 		
 		hitBox = new Rectangle2D.Float(x, y, widht, height);
